Move ViewTransitions helpers to ReactFiberCommitViewTransitions
